Extract logging handling to AbstractIntegrationTest
